Deciding how to handle the dependency version matrix (#5)
This was discussed in the 2024-03-14 office hours and no blockers were brought up. We recognize that this RFC makes trade-offs in a complicated space and there will be downsides. Even so, the decision is to move forward with its implementation. Further refinements can always come in the future as we learn from experience and the tool ecosystem improves.